add multiple inference optimization for CorrDiff

+++ physicsnemo/models/diffusion/song_unet.py
@@ -419,7 +419,9 @@ class SongUNet(Module):
                 emb = silu(self.map_layer1(emb))
             else:
                 emb = torch.zeros(
-                    (noise_labels.shape[0], self.emb_channels), device=x.device
+                    (noise_labels.shape[0], self.emb_channels),
+                    device=x.device,
+                    dtype=x.dtype,
                 )
 
             # Encoder.
